backend/requirement.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0
requests>=2.31
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.6
alembic>=1.13
pydantic>=2.0
cryptography>=41.0
